Array Structure combining lecture and adding notes
Combining collection expressions into the lecture as it makes sense to do so, declaration and initialization are core part of the array structure so talking about it separately didn't make sense. Spread operator will be kept separate as it is an extended functionality. Adding time complexities moving forward after CS fundamentals covered.