Multidimensional Array content created, reorg
Creating an example of a treasure map - gets iterated, conditions checked if X is surrounded by stars (*) up, down, left and right it is a hidden treasure otherwise it is not a hidden treasure. Demonstrates realistic scenario of bound checking and 2d array iteration.

Also reorganizing multidimensional arrays and jagged arrays.