Removing Sealed Modifier and reorganizing
Sealed Modifier example does not make much sense as it should mostly be favored not to use sealed to improving testability of the system. It is needed in niche situations which could be more confusing than helpful. If it is requested by students to be added then based on the demand it will be added. Initial decision is to exclude it.