Structure & Iteration will be combined
The reason for this is that during Arrays lecture these were separately explained in detail. Strings are very similar to Arrays and so I can skip other loops and explain for loop only and combine it all into Structure. If I need to separate it in the future that is OK.